Add fallback to eth1 providerUrls
