+++ packages/lodestar/src/eth1/index.ts
@@ -4,6 +4,6 @@
 
 export * from "./interface";
 export * from "./stream";
-export * from "./eth1Provider";
+export * from "./provider/eth1Provider";
 export * from "./eth1ForBlockProduction";
 export * from "./eth1ForBlockProductionDisabled";
