+++ packages/lodestar/src/eth1/interface.ts
@@ -9,11 +9,11 @@ import {CachedBeaconState} from "@chainsafe/lodestar-beacon-state-transition";
 
 export interface IEth1Provider {
   deployBlock: number;
-  getBlockNumber(signal?: AbortSignal): Promise<number>;
-  getBlockByNumber(blockNumber: number, signal?: AbortSignal): Promise<phase0.Eth1Block>;
-  getBlocksByNumber(fromBlock: number, toBlock: number, signal?: AbortSignal): Promise<phase0.Eth1Block[]>;
-  getDepositEvents(fromBlock: number, toBlock: number, signal?: AbortSignal): Promise<phase0.DepositEvent[]>;
-  validateContract(signal?: AbortSignal): Promise<void>;
+  getBlockNumber(): Promise<number>;
+  getBlockByNumber(blockNumber: number): Promise<phase0.Eth1Block>;
+  getBlocksByNumber(fromBlock: number, toBlock: number): Promise<phase0.Eth1Block[]>;
+  getDepositEvents(fromBlock: number, toBlock: number): Promise<phase0.DepositEvent[]>;
+  validateContract(): Promise<void>;
 }
 
 export interface IEth1ForBlockProduction {
@@ -49,7 +49,11 @@ export interface IRpcPayload {
   params: IJson[];
 }
 
+export type ReqOpts = {
+  timeout?: number;
+};
+
 export interface IJsonRpcClient {
-  fetch<R>({method, params}: IRpcPayload, signal?: AbortSignal): Promise<R>;
-  fetchBatch<R>(rpcPayloadArr: IRpcPayload[], signal?: AbortSignal): Promise<R[]>;
+  fetch<R>({method, params}: IRpcPayload, opts?: ReqOpts): Promise<R>;
+  fetchBatch<R>(rpcPayloadArr: IRpcPayload[], opts?: ReqOpts): Promise<R[]>;
 }
