+++ packages/lodestar/src/eth1/options.ts
@@ -1,11 +1,11 @@
 export interface IEth1Options {
   enabled: boolean;
-  providerUrl: string;
+  providerUrls: string[];
   depositContractDeployBlock: number;
 }
 
 export const defaultEth1Options: IEth1Options = {
   enabled: true,
-  providerUrl: "http://localhost:8545",
+  providerUrls: ["http://localhost:8545"],
   depositContractDeployBlock: 0,
 };
